refactor/#61: refactoring coordinator pattern logic
